test: migrate StudyAssetList to vitest
